requirements.txt
GitPython
PyYAML
Pillow
py-robocopy
